+++ src/app/test-chaemin/page.tsx
@@ -3,7 +3,7 @@ import CommentBox from "@component/components/comment/CommentBox";
 import Button from "@component/components/common-components/button";
 import Modal from "@component/components/common-components/modal";
 import Tag from "@component/components/common-components/tag";
-import { Header } from "@component/components/common/Header";
+import { Header } from "@component/components/common-components/common/Header";
 import ProjectItem from "@component/components/landing/project/ProjectItem";
 import InterestModal from "@component/components/sign-up/InterestModal";
 import LoginModal from "@component/components/sign-up/LoginModal";
