+++ src/app/test/page.tsx
@@ -9,7 +9,7 @@ import TabComponent from "@component/components/common-components/tab/TabCompone
 import LandingTop from "@component/components/landing/carousel/LandingMid";
 import Button from "@component/components/common-components/button/Button";
 import Dropdown from "@component/components/common-components/dropdown";
-import Footer from "@component/components/footer/Footer";
+import Footer from "@component/components/common-components/common/Footer";
 
 const Test = () => {
   // input
@@ -283,7 +283,6 @@ const Test = () => {
         className={isInvalid ? "border-error-main" : "border-purple-main1"}
       />
       <Button onClick={focus}>제출하기</Button>
-
       <Footer />
     </div>
   );
